Separate the Persona class from the test code
